refactor: extract SessionInterface and listAll to api_helper.go
- Moved SessionInterface and its implementation to a dedicated file for better code organization
- Extracted listAll helper function to api_helper.go as it's a utility function related to the SessionInterface
- No functional changes, only code reorganization
